requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
requests==2.26.0
numpy==1.21.2
pandas==1.3.3
scipy==1.7.1
websocket-client==1.2.1
python-dotenv==0.19.0
eventlet==0.33.0
PyJWT==2.1.0
ta==0.7.0
websockets==9.1
python-telegram-bot==13.7 